Refactor test workflows and improve logging: update test commands, enhance logging configuration, and clean up unused files.
