+++ .github/workflows/deploy.yml
@@ -26,7 +26,8 @@ jobs:
           pip install pytest pytest-asyncio httpx
 
       - name: Run tests
-        run: pytest -v --disable-warnings
+        # run: pytest -v --disable-warnings
+        run: pytest -v tests/test_users.py --disable-warnings
 
   deploy:
     name: Deploy to Server
