--- app/visits.txt
@@ -1,1 +1,0 @@
-31
