+++ requirements-dev.txt
@@ -1,4 +1,4 @@
-asgi-lifespan==2.1.0
-respx==0.22.0
+asgi_lifespan==2.1.0
 pytest==8.3.5
-pytest-asyncio==1.0.0
+pytest_asyncio==1.0.0
+respx==0.22.0
