+++ requirements.txt
@@ -1,46 +1,31 @@
+aioredis==1.3.1
 annotated-types==0.7.0
 anyio==4.9.0
+async-timeout==5.0.1
 certifi==2025.4.26
-click==8.1.8
+click==8.2.1
 colorama==0.4.6
-dnspython==2.7.0
-email_validator==2.2.0
-exceptiongroup==1.2.2
+dotenv==0.9.9
+exceptiongroup==1.3.0
 fastapi==0.115.12
 fastapi-cache2==0.2.2
-fastapi-cli==0.0.7
 h11==0.16.0
+hiredis==3.2.1
 httpcore==1.0.9
-httptools==0.6.4
 httpx==0.28.1
 idna==3.10
-itsdangerous==2.2.0
 Jinja2==3.1.6
-markdown-it-py==3.0.0
 MarkupSafe==3.0.2
-mdurl==0.1.2
-orjson==3.10.18
 pendulum==3.1.0
-pydantic==2.11.4
-pydantic-extra-types==2.10.4
-pydantic-settings==2.9.1
+pydantic==2.11.5
 pydantic_core==2.33.2
-Pygments==2.19.1
 python-dateutil==2.9.0.post0
 python-dotenv==1.1.0
 python-multipart==0.0.20
-PyYAML==6.0.2
-rich==14.0.0
-rich-toolkit==0.14.6
-shellingham==1.5.4
 six==1.17.0
 sniffio==1.3.1
 starlette==0.46.2
-typer==0.15.4
-typing-inspection==0.4.0
+typing-inspection==0.4.1
 typing_extensions==4.13.2
 tzdata==2025.2
-ujson==5.10.0
 uvicorn==0.34.2
-watchfiles==1.0.5
-websockets==15.0.1
