+++ service/cache.py
@@ -1,8 +1,10 @@
 # service/cache.py
-from fastapi_cache import FastAPICache
-from datetime import datetime, timedelta, timezone
 import logging
-from service.config import CACHE_TTL
+from datetime import datetime, timedelta, timezone
+
+from fastapi_cache import FastAPICache
+
+from service.configs import CACHE_TTL
 
 
 def get_backend():
@@ -22,25 +24,17 @@ async def set_cached(key: str, value: dict, ttl: int):
     await backend.set(key, value, expire=ttl)
 
 
-def calculate_next_update(last_update_iso: str) -> datetime:
-    last_update = datetime.fromisoformat(last_update_iso.replace("Z", "+00:00"))
-    minute = (last_update.minute // 15 + 1) * 15
-    if minute == 60:
-        next_update = last_update.replace(
-            minute=0, second=0, microsecond=0
-        ) + timedelta(hours=1)
-    else:
-        next_update = last_update.replace(minute=minute, second=0, microsecond=0)
-    return next_update.astimezone(timezone.utc)
-
-
 def ttl_logic(
     data: dict,
     source: str = "auto",
     return_ttl: bool = False,
-    fallback_ttl: int | None = None,
 ) -> int | bool:
+    """
+    Для любого source возвращает True если кэш валиден (до конца текущего интервала),
+    либо TTL до конца интервала (интервал задаётся через CACHE_TTL в секундах, если return_ttl=True).
+    """
     try:
+        # Определяем ключ для CACHE_TTL
         if source == "auto":
             if "current_weather" in data:
                 source = "weather"
@@ -49,38 +43,27 @@ def ttl_logic(
             ):
                 source = "cat"
             else:
-                raise ValueError("Неизвестный тип данных")
+                source = None
 
-        if source == "weather":
-            now_utc = datetime.now(timezone.utc)
-            minute = (now_utc.minute // 15) * 15
-            interval_start = now_utc.replace(minute=minute, second=0, microsecond=0)
-            interval_end = interval_start + timedelta(minutes=15)
+        # Определяем имя ключа для CACHE_TTL
+        cache_key = f"{source}_cache" if source else None
+        interval_sec = CACHE_TTL.get(cache_key, 60)
+        now_utc = datetime.now(timezone.utc)
+        # Вычисляем начало и конец текущего интервала
+        interval_start = now_utc - timedelta(seconds=now_utc.timestamp() % interval_sec)
+        interval_end = interval_start + timedelta(seconds=interval_sec)
 
-            if return_ttl:
-                # Используем CACHE_TTL если есть, иначе считаем по времени
-                ttl_sec = CACHE_TTL.get(
-                    "weather_cache", int((interval_end - now_utc).total_seconds())
-                )
-                return max(0, int(ttl_sec))
+        if return_ttl:
+            # TTL до конца текущего интервала
+            ttl = int((interval_end - now_utc).total_seconds())
+            # Если вдруг ttl=0, то возвращаем полный интервал, чтобы не было нулевого TTL
+            return ttl if ttl > 0 else interval_sec
 
-            return now_utc < interval_end
-
-        if source == "cat":
-            if return_ttl:
-                # Используем CACHE_TTL если есть, иначе fallback_ttl или 60
-                return CACHE_TTL.get(
-                    "cat_cache", fallback_ttl if fallback_ttl is not None else 60
-                )
-            return True
+        # Кэш валиден, если текущее время меньше конца интервала
+        return now_utc < interval_end
 
     except Exception as e:
         logging.warning(f"TTL error for {source}: {e}")
         if return_ttl:
-            # Используем fallback_ttl если есть, иначе значение из CACHE_TTL или 60
-            return (
-                fallback_ttl
-                if fallback_ttl is not None
-                else CACHE_TTL.get("cat_cache", 60)
-            )
+            return 60
         return False
