+++ service/variables.py
@@ -1,5 +1,5 @@
-from pathlib import Path
 from datetime import datetime
+from pathlib import Path
 
 # Константы приложения
 BASE_DIR = Path(__file__).resolve().parent.parent
