+++ tests/pytest.ini
@@ -2,10 +2,6 @@
 asyncio_mode = strict
 asyncio_default_fixture_loop_scope = module
 
-markers =
-    cat: marker for cat tests
-
-
 log_format = %(levelname)-8s | %(message)s | %(asctime)s
 log_date_format = %Y-%m-%d %H:%M:%S
 log_cli = true
