+++ tests/test_cat.py
@@ -2,13 +2,13 @@
 import pytest
 import respx
 from httpx import Response
+
 from service.variables import CAT_FALLBACK
 
 
 # 1. Успешный случай с моками
 @respx.mock
 @pytest.mark.asyncio
-@pytest.mark.category("cat")
 async def test_api_cat_success(client):
     mock_url = "https://cdn.fakecat.com/cat.jpg"
     api_mock = respx.get("https://api.thecatapi.com/v1/images/search").mock(
