+++ tests/test_main.py
@@ -1,6 +1,7 @@
 # tests/test_main.py
 from fastapi.testclient import TestClient
-from main import app  # путь подкорректируй под свою структуру
+
+from main import app
 
 client = TestClient(app)
